Convert homepage to standalone router
